Replace Euler angle usage in geometry.points_on_1sphere
A test shows that euler_to_quaternion(0, theta, 0) and angle_axis_to_rot3d('y', theta) are equivalent.
A test shows that euler_to_quaternion(0, 0, phi) and angle_axis_to_rot3d('x', phi) are equivalent.
The former is therefore replaced by the later in the points_on_1sphere function.
It is worth noting that, in the latter case, points_on_1sphere wrongly thought it was acting on the 'z' axis.